# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Layth and Rashid",
      "version": "1.0.0",
      "title": "CourseConnects",
      "host":	"courseconnects.onrender.com",
      "basePath":"/v2",
      "contact": {
        "email": "seanwasere@gmail.com"
      }
    },
    "servers": [
      {
        "url": "/"
      }
    ],
...
    "paths": {
...
              "/api/addfriend": 
              {
                  "post": {
                    "tags": [
                      "Friends"
                    ],
                    "summary": "add a friend",
                    
                    "produces": [
                      "application/json"
                    ],
                    "responses": {
                      "201": {
                        "description": "OK"
                      },
                      "400": {
                        "description": "Failed. Bad post data."
                      }
                    },
                    "requestBody": {
                      "description": "Add Friend Request Post Object",
                      "required": true,
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/AddFriendRequestPostBody"
                          }
                        }
                      }
                    }
                  }
                },
                "/api/getfriends": 
                {
                    "get": {
                      "tags": [
                        "Friends"
                      ],
                      "summary": "get your friends",
                      
                      "produces": [
                        "application/json"
                      ],
                      "responses": {
                        "201": {
                          "description": "OK"
                        },
                        "400": {
                          "description": "Failed. Bad post data."
                        }
                      }
                      
                    }
                  },
...
          },
    "components": {
      "schemas": {
...
            "AddFriendRequestPostBody": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "format": "string"
                }
              }
            },
...
        }
      }
    }